requirements.txt
django==4.1.4
celery==5.2.7
django-filter==22.1
social-auth-app-django==5.0.0
django-modeltranslation==0.18.7
djangorestframework==3.14.0
django-rest-knox==4.2.0
django-celery-beat==2.4.0
psycopg2==2.9.5
django-multiselectfield==0.1.12
djangorestframework-simplejwt==5.2.2
pillow==9.3.0
